+++ init.lua
@@ -1,8 +1,2 @@
-require("user.nvim-options")
-require("user.autocommands")
-require("user.mappings")
-require("user.colorscheme")
-
-require("plugins")
-require("user.utils")
-require("user.lsp")
+require("mengano.core")
+require("mengano.lazy")